+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+### Added
+- `pw.Table.filter_out_results_of_forgetting` method, allowing to revert the effects of forgetting at a later stage.
+
 ## [0.27.0] - 2025-11-13
 
 ### Added